backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
snowflake-connector-python==3.5.0
python-dotenv==1.0.0
